01_Active_Projects/memory_agent/requirements.txt
# 学习记忆助手Agent 依赖

# 向量数据库
chromadb>=0.4.22

# 文本嵌入（本地运行）
sentence-transformers>=2.3.1

# 文本处理
jieba>=0.42.1

# 配置管理
pyyaml>=6.0.1

# 进度条
tqdm>=4.66.1

# 日期时间处理
python-dateutil>=2.8.2
